Move legislature model
