+++ src/parliament/mappers/legislature.py
@@ -1,8 +1,10 @@
-from parliament.model import Legislatura
+from parliament.models.legislature import Legislatura
 
 
 def map_to_legislature(leg: str) -> Legislatura:
     match leg:
+        case "XIV":
+            return Legislatura.XIV
         case "XV":
             return Legislatura.XV
         case "XVI":
