+++ src/parliament/models/agenda.py
@@ -1,6 +1,6 @@
 from enum import Enum
 from typing import Optional
-from parliament.model import Legislatura
+from parliament.models.legislature import Legislatura
 from pydantic import BaseModel
