+++ src/parliament/models/legislature.py
@@ -2,5 +2,6 @@ from enum import Enum
 
 
 class Legislatura(str, Enum):
+    XIV = "XIV"
     XV = "XV"
     XVI = "XVI"
