+++ src/parliament/routers/agenda.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, HTTPException
 from httpx import AsyncClient
 from parliament.mappers.agenda import map_to_upcoming_events
 from parliament.models.agenda import EventoAgenda
-from parliament.model import Legislatura
+from parliament.models.legislature import Legislatura
 from parliament.constants import current_legislature
